backend/requirements.txt

fastapi
uvicorn[standard]
pymongo
sentence-transformers
scikit-learn
requests
python-multipart
pymupdf
python-docx
openpyxl
extract-msg
python-pptx
beautifulsoup4
python-magic
apscheduler
joblib
fitz
